requirements.txt
requests==2.28.1
python_telegram_bot==20.0
requests_cache==0.9.7